Update Exercise5.md
I got a little lost jumping from Excercise 5 to 6 because I didn't see an entire new section (which is how I'm used to thinking of trees). It took me several minutes and some clicking around to find it and only then did I realize that the code told me where it was. 

My goal here is just to highlight that part of the code and explicitly introduce users to the location of the ingredient tree before the next exercise when you're instructed to click on it. It doesn't have to use this language of course.